Include store_dataset_from_ddf in shuffle tests
